Add authorization guard and refactor event application
Event application and preference creation and getting have been secured by associating them with a user account. Service methods for these have been changed to create or fetch based on user account. Event application pair creation has been updated to either update an existing pair or create a new one depending on whether a suitable pair exists. Preference resolver has been updated to have delete capability, and authorization has also been added to ensure only the owner can delete a preference. Group service has been updated to fetch groups based on account id. Authorization guards have been added to resolvers to check for access rights.

+++ src/app/event-application/preference/preference.resolver.ts
@@ -8,13 +8,16 @@ import {
   ResolveField,
   Resolver,
 } from '@nestjs/graphql';
+import { I18nService } from 'nestjs-i18n';
 
 import { EventApplication } from '@/@generated/nestgraphql/event-application/event-application.model';
 import { Preference } from '@/@generated/nestgraphql/preference/preference.model';
 import { PriceRange } from '@/@generated/nestgraphql/prisma/price-range.enum';
 import { AuthGuard } from '@/app/auth/auth-guard/auth.guard';
+import { RequestContext } from '@/app/auth/request-context-extractor/interfaces';
 import { EventApplicationService } from '@/app/event-application/event-application.service';
 import { PreferenceService } from '@/app/event-application/preference/preference.service';
+import { RequestContextDecorator } from '@/app/request-context.decorator';
 
 @InputType()
 export class CreatePreferenceInput {
@@ -39,19 +42,53 @@ export class PreferenceResolver {
   constructor(
     private readonly preferenceService: PreferenceService,
     private readonly eventApplicationService: EventApplicationService,
+    private i18n: I18nService,
   ) {}
 
-  @Mutation(() => Preference, { name: 'createPreference' })
-  @UseGuards(AuthGuard)
-  async createPreference(
-    @Args('input') input: CreatePreferenceInput,
-  ): Promise<Preference> {
-    return this.preferenceService.createPreference(input);
+  private async authorize(
+    context: RequestContext,
+    preference: Preference,
+  ): Promise<void> {
+    const applications =
+      await this.eventApplicationService.getEventApplicationByAccountId(
+        context.account!.id,
+      );
+
+    if (!applications) {
+      // eslint-disable-next-line sonarjs/no-duplicate-string
+      throw new Error(this.i18n.t('errors.unauthorized'));
+    }
+
+    const eventApplication =
+      await this.eventApplicationService.getEventApplicationByPreferenceId(
+        preference.id,
+      );
+
+    if (!eventApplication) {
+      // eslint-disable-next-line sonarjs/no-duplicate-string
+      throw new Error(this.i18n.t('errors.unauthorized'));
+    }
+
+    const isApplicationOwner = applications.some(
+      (application) => application.id === eventApplication.accountId,
+    );
+
+    if (!isApplicationOwner) {
+      throw new Error(this.i18n.t('errors.unauthorized'));
+    }
   }
 
   @Mutation(() => Preference, { name: 'deletePreference' })
   @UseGuards(AuthGuard)
-  async deletePreference(@Args('id') id: number): Promise<Preference> {
+  async deletePreference(
+    @Args('id') id: number,
+    @RequestContextDecorator() context: RequestContext,
+  ): Promise<Preference> {
+    const preference = await this.preferenceService.getPreferenceById(id);
+    if (!preference) {
+      throw new Error(this.i18n.t('errors.notFound'));
+    }
+    await this.authorize(context, preference);
     return this.preferenceService.deletePreference(id);
   }
 
@@ -59,7 +96,9 @@ export class PreferenceResolver {
   @UseGuards(AuthGuard)
   async eventApplication(
     @Parent() preference: Preference,
+    @RequestContextDecorator() context: RequestContext,
   ): Promise<EventApplication | null> {
+    await this.authorize(context, preference);
     return this.eventApplicationService.getEventApplicationByPreferenceId(
       preference.id,
     );
